% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_features.R
\name{add_clustering}
\alias{add_clustering}
\title{Add feature clustering}
\usage{
add_clustering(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\description{
Cluster features on t-values
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   object <- subramanian.2016::exiqon
   object \%>\% autonomics.plot::plot_overlayed_sample_distributions()
}
if require(autonomics.data){
    object <- autonomics.data::glutaminase
    object$TREATMENT <- stringi::stri_split_fixed(object$subgroup, '_') \%>\% 
    vapply(extract, character(1), 1)
    group_var <- 'TREATMENT'
    object \%<>\% add_clustering()
}
}
