% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_limma.R
\name{add_limma}
\alias{add_limma}
\title{Run limma and add results}
\usage{
add_limma(object,
  contrastdefs = autonomics.find::default_contrasts(object),
  design = autonomics.import::create_design_matrix(object))
}
\arguments{
\item{object}{SummarizedExperiment}

\item{contrastdefs}{contrastdef vector, preferably named (automatically generated names are not always intuitive)}

\item{design}{design matrix}
}
\value{
Updated SummarizedExperiment
}
\description{
Limma results can be easily accessed with limma(object).
This returns a list with components:
\itemize{
   \item {effect} matrix (ngene x ncontrast): effect sizes
   \item {rank}   matrix (ngene x ncontrast): effect ranks
   \item {t}      matrix (ngene x ncontrast): t    values (moderated t test)
   \item {se}     matrix (ngene x ncontrast): se   values (moderated t test)
   \item {p}      matrix (ngene x ncontrast): p    values (moderated t test)
   \item {fdr}    matrix (ngene x ncontrast): fdr  values (moderated t test)
   \item {bonf}   matrix (ngene x ncontrast): bonf values (moderated t test)
   \item {F}      vector (ngene)            : F    values (moderated F test)
   \item {F.p}    vector (ngene)            : p    values (moderated F test)
}
}
\examples{
require(magrittr)
#if (require(atkin.2014)){
#   results <- atkin.2014::soma  \%>\%  autonomics.find::run_limma()
#}

# STEM CELL DIFFERENTIATION
if (require(autonomics.data)){
   object <- autonomics.data::stemdiff.proteinratios
   contrastdefs <- autonomics.find::make_ref_contrasts(object)
   object \%>\% autonomics.find::add_limma(contrastdefs[1:2]) \%>\% 
              extract2('bonf')       \%>\%
              extract(, 'EM01_EM00') \%>\% 
              is_less_than(0.05)     \%>\% 
              sum(na.rm=TRUE)

}
if (require(billing.differentiation.data)){
   contrastdefs <- billing.differentiation.data::contrasts

   billing.differentiation.data::rna.counts \%>\%
      autonomics.import::logcpm() \%>\%
      autonomics.find::run_limma(contrastdefs) \%>\% extract2('bonf') \%>\%
      extract(, 'EM01_0') \%>\% is_less_than(0.05) \%>\% sum(na.rm=TRUE)

   billing.differentiation.data::rna.counts \%>\%
      autonomics.import::voom_transform() \%>\%
      autonomics.find::run_limma(contrastdefs) \%>\% extract2('bonf') \%>\%
      extract(, 'EM01_0') \%>\% is_less_than(0.05) \%>\% sum(na.rm=TRUE)
}
if (require(subramanian.2016)){
   dir <- system.file('extdata/rnaseq', package = 'subramanian.2016')
   object <- autonomics.import::load_rnaseq_counts(dir, sample_file = NULL) \%>\% 
             autonomics.import::voom_transform()
   results <- object \%>\% autonomics.find::run_limma()
   results$p[1:3,1:3]
}
if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon
}

}
