% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{aggregate_contrasts}
\alias{aggregate_contrasts}
\title{Aggregate time contrasts across conc (or vice versa)}
\usage{
aggregate_contrasts(contrastmat, dim)
}
\arguments{
\item{contrastmat}{contrast matrix}
}
\description{
Aggregate time contrasts across conc (or vice versa)
}
\examples{
(x <- create_subgroup_matrix(autonomics.data::glutaminase))
aggregate_contrasts(create_conc_contrasts(x), 1) # some conc across t
aggregate_contrasts(create_conc_contrasts(x), 2) # concentrations at t
aggregate_contrasts(create_time_contrasts(x), 2) # some time across conc
aggregate_contrasts(create_time_contrasts(x), 1) # times at conc
}
