% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_contrast_features.R
\name{are_top_features}
\alias{are_top_features}
\alias{are_contrast_features}
\alias{get_top_features}
\alias{get_contrast_features}
\title{Are contrast features?}
\usage{
are_top_features(object, topdef, contrast_name)

are_contrast_features(
  object,
  contrast_name,
  topdef = autonomics.find::default_topdef(object)
)

get_top_features(object, topdef, contrast_name)

get_contrast_features(
  object,
  contrast_name,
  topdef = autonomics.find::default_topdef(object)
)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{topdef}{character(1): top definition}

\item{contrast_name}{character(1): contrast name}

\item{...}{for backward compatibility only}
}
\value{
integer vector with indices of top features
}
\description{
Are contrast features?
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::stemdiff.proteinratios
   topdef <- 'p < 0.05'
   contrast_name <- object \%>\% autonomics.import::contrastdefs() \%>\% magrittr::extract(1) \%>\% names()
   object \%>\% autonomics.find::are_contrast_features('p<0.05',            contrast_name) \%>\% sum()
   object \%>\% autonomics.find::are_contrast_features('p<0.05 & effect>0', contrast_name) \%>\% sum()
}
}
