% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_relevant_samples.R
\name{are_relevant_samples}
\alias{are_relevant_samples}
\title{Are relevant samples for given contrast}
\usage{
are_relevant_samples(
  object,
  design = autonomics.import::create_design_matrix(object),
  contrasts
)
}
\arguments{
\item{object}{eset}

\item{design}{design matrix}

\item{contrasts}{contrasts}
}
\description{
Are relevant samples for given contrast
}
\examples{
require(magrittr)

if (require(autonomics.data)){
   object <- autonomics.data::ALL
   unique(object$subgroup)
   contrasts <- c(M_F = '(TM+BM)/2 - (TF+BF)/2', B_T = '(BM+BF)/2 - (TM+TF)/2')
   object \%<>\% autonomics.find::add_limma_to_fdata(contrasts = contrasts)
   object \%>\%  autonomics.find::are_relevant_samples(contrasts = contrasts)
}

# STEM CELL COMPARISON
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios
   contrasts <- object$subgroup[1]
   object \%>\% autonomics.find::are_relevant_samples(contrasts = contrasts)
}

if (require(billing.differentiation.data)){
   object <- billing.differentiation.data::rna.voomcounts
   object \%>\% dim()
   contrasts <-  c(EM0.8_0 = 'EM0.8 - EM0.0')
   object \%>\% autonomics.find::are_relevant_samples(   contrasts = contrasts)
}
}
