% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_n_arrange_top_features.R
\name{are_top_features}
\alias{are_top_features}
\alias{get_top_features}
\title{Which are the top features?}
\usage{
are_top_features(object, topdef = autonomics.find::default_topdef(object),
  contrast_name, direction)

get_top_features(object, topdef = autonomics.find::default_topdef(object),
  contrast_name, direction)
}
\arguments{
\item{object}{eset}

\item{topdef}{top definition}

\item{contrast_name}{contrast name (string)}

\item{direction}{'both', 'pos', or 'neg'}
}
\value{
integer vector with indices of top features
}
\description{
Which features fulfill top definition and direction for specified contrast?
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   object <- billing.differentiation.data::rna.voomcounts
   contrast <-  c(EM0.8_0 = 'EM0.8 - EM0.0')
   topdef <- 'bonf < 0.05'
   autonomics.find::are_top_features(
      object, topdef, names(contrast), direction = 'neg')  \%>\% 
      sum()
   autonomics.find::are_top_features(
      object, topdef, names(contrast), direction = 'pos')  \%>\% 
      sum()
   autonomics.find::are_top_features(
      object, topdef, names(contrast), direction = 'both') \%>\% 
      sum()
}
}
