% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_n_arrange_top_features.R
\name{arrange_features_by_rank}
\alias{arrange_features_by_rank}
\title{Order features on rank}
\usage{
arrange_features_by_rank(object, contrast_name)
}
\arguments{
\item{object}{eSet}

\item{contrast_name}{contrast name}
}
\value{
arranged eSet
}
\description{
Order features on rank
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   object <- billing.differentiation.data::rna.voomcounts
   object \%>\% autonomics.import::fnames() \%>\% head()
   contrast <-  c(EM0.8_0 = 'EM0.8 - EM0.0')
   object \%>\% autonomics.find::arrange_features_by_rank(names(contrast)) \%>\% 
                autonomics.import::fnames() \%>\% head()
}
}
