% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_is_valid_limma_contrast}
\alias{assert_is_valid_limma_contrast}
\title{Assert validity of limma contrast}
\usage{
assert_is_valid_limma_contrast(contrast, an_eset)
}
\arguments{
\item{contrast}{limma contrast (character)}

\item{an_eset}{eSet}
}
\description{
Assert validity of limma contrast
}
