% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigAB.R
\name{calc_sig_a}
\alias{calc_sig_a}
\title{Calculate sig A}
\usage{
calc_sig_a(x, force = FALSE, method = "none")
}
\arguments{
\item{x}{Vector of numerical values to calculate the outlier significance for 
(see "Details" for more).}

\item{force}{Boolean indicating whether or not to go ahead with likely 
non-logarithmized data (see "Details" for more).}

\item{method}{Method for adjusting p-values for multiple comparisons. See
\code{\link[stats]{p.adjust}}.}
}
\value{
Returns a vector of p-Values.
}
\description{
A method to calculate significance A as described in the original MaxQuant
publication (with modifications)
}
\details{
The numerical values fed to this function should represent an approx.
symetrical distribution and might require prior logarithmization.

Details of the calculation may be looked up in the supplementary material of
the citation listed (page 19f).
The algorithm differs from that original - emulating later MaxQuant/Perseus
versions - in multiplying p-values by a factor of two, thus converting the
originally one-sided test into a two-sided one (J. Cox, personal communication).

\code{force=FALSE} implies that the data fed to the algorithm is checked for 
prior logarithmization. The function throws an error if \code{min(x,na.rm=TRUE) >= 0}. 
This behavior may be overridden by setting \code{force=TRUE}.

This implementation shows good correlation with the one performed by
Perseus as of version 1.1.1.17. See the example below for a
demonstration.

Note that no multiple hypothesis testing is done.
}
\references{
Cox, J., and Mann, M. (2008). MaxQuant enables high peptide identification
rates, individualized p.p.b.-range mass accuracies and proteome-wide protein
quantification. Nat. Biotechnol 26, 1367-1372.
This function is borrowed from pdapmain; see that pkg for examples.
}
\seealso{
\code{\link{calc_sig_b}}, \code{\link[stats]{p.adjust}}
}
\author{
Johannes Graumann
}
\keyword{distribution}
\keyword{manip}
\keyword{methods}
