% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigAB.R
\name{calc_sig_b}
\alias{calc_sig_b}
\title{Calculate sig B}
\usage{
calc_sig_b(x, y, BinSize = 500, BinType = c("Hard", "Optimal"), ...)
}
\arguments{
\item{x}{A vector of numerical values to calculate the outlier significance for. See 
"Details" for more.}

\item{y}{A vector of numerical values according to which the binning of values in 
\code{x} #' will be done. See "Details"for more.}

\item{BinSize}{The minimum number of values per bin from \code{x} binned according to 
\code{y}. Defaulting to 500.}

\item{BinType}{Mode of bin positioning and size determination (see "Details"; 
defaulting to "Hard").}

\item{...}{Passed to \code{\link{calc_sig_a}}.}
}
\value{
Returns a vector of p-Values.
}
\description{
A method to calculate significance B as described in the original MaxQuant
publication (with modifications)
}
\details{
Significance B equates to \link[=calc_sig_a]{Significance A}
on data that has bin binned according to a binning vector "y" (summed peptide intensity 
in the original application) and a minimum bin size.

The numerical values fed to this function as "x" should represent an approx. symetrical 
distribution and might require prior logarithmization.

Details of the calculation may be looked up in the supplementary material of the 
citation listed (page 19f).

The algorithm differs from that original - emulating later MaxQuant/Perseus versions - 
by using a default bin size of 500 (J. Cox, personal communication).

\code{BinType} indicates how values should be distributed if \code{length(x)} is not a 
multiple of \code{BinSize}. When using the MaxQuant/Perseus-analogous 
\code{BinType="Hard"}, all but the last bin contain \code{BinSize} members, while 
\code{BinType="Optimal"} calculates the maximum number of bins with at least 
\code{BinSize} members and evenly distributes the members to those bins.

See the example below for a comparative correlation with the one performed by Perseus 
as of version 1.1.1.17.

Note that no multiple hypothesis testing is done.
}
\references{
Cox, J., and Mann, M. (2008). MaxQuant enables high peptide identification
rates, individualized p.p.b.-range mass accuracies and proteome-wide protein
quantification. Nat. Biotechnol 26, 1367-1372.
This function is borrowed from pdapmain; see that pkg for examples.
}
\seealso{
\code{\link{calc_sig_a}}
}
\author{
Johannes Graumann
}
\keyword{distribution}
\keyword{manip}
\keyword{methods}
