% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_features.R
\name{cluster_features}
\alias{cluster_features}
\title{Cluster features}
\usage{
cluster_features(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\description{
Cluster features
}
\examples{
require(magrittr)

if (require(autonomics.data)){

   # Stem Cell Comparison (LCMSPROT)
   #--------------------------------
   object <- autonomics.data::stemcomp.proteinratios
   
   # Stem Cell differentiation (LCMSPROT)
   #-------------------------------------
   object <- autonomics.data::stemdiff.proteinratios

   # Glutaminase inhibitor (METABOLON)
   #----------------------------------
   object <- autonomics.data::glutaminase
   
}
}
