% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_top_features.R
\name{cluster_features_on_subgroups}
\alias{cluster_features_on_subgroups}
\title{Cluster features on subgroup profiles}
\usage{
cluster_features_on_subgroups(
  object,
  result_dir,
  geom = autonomics.plot::default_feature_plots(object) \%>\% setdiff("bars"),
  x = autonomics.plot::default_x(object, geom[1]),
  color_var = autonomics.plot::default_color_var(object),
  shape_var = autonomics.plot::default_shape_var(object),
  group_var = autonomics.plot::default_group_var(object),
  txt_var = autonomics.plot::default_txt_var(object),
  line = TRUE
)
}
\arguments{
\item{object}{eset}

\item{result_dir}{result directory}

\item{geom}{which types of feature plots to generate}

\item{x}{svar mapped to x     in feature plots}

\item{color_var}{svar mapped to color in feature plots}

\item{shape_var}{svar mapped to shape in feature plots}

\item{group_var}{svar mapped to group in feature plots}

\item{txt_var}{svar mapped to txt   in feature plots}

\item{line}{whether to connect points with a line in feature plots}
}
\description{
Cluster features on subgroup profiles
}
\examples{
\dontrun{
   require(magrittr)
   result_dir <- tempdir() \%T>\% message()
   if (require(subramanian.2016)){
      object <- subramanian.2016::metabolon[1:100,]
      object \%>\% cluster_features_on_subgroups(
                      result_dir, color_var = 'condition', group_var = 'condition')
   }
}
}
