% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_top_features.R
\name{cluster_top_features_on_subgroups}
\alias{cluster_top_features_on_subgroups}
\title{Cluster top features (on subgroup profiles)}
\usage{
cluster_top_features_on_subgroups(
  object,
  design = autonomics.import::create_design_matrix(object),
  contrasts = autonomics.find::default_contrasts(object),
  result_dir,
  n = 1000,
  geom = autonomics.plot::default_feature_plots(object) \%>\% setdiff("bars"),
  x = autonomics.plot::default_x(object, geom[1]),
  color_var = autonomics.plot::default_color_var(object),
  shape_var = autonomics.plot::default_shape_var(object),
  group_var = autonomics.plot::default_group_var(object),
  txt_var = autonomics.plot::default_txt_var(object),
  line = TRUE
)
}
\arguments{
\item{object}{eSet updated with add_limma_to_fdata}

\item{design}{design matrix}

\item{contrasts}{contrasts definition}

\item{result_dir}{result directory}

\item{n}{number of top features}

\item{geom}{which type of feature plots to generate}

\item{x}{svar mapped to x in feature plots}

\item{color_var}{svar mapped to color in feature plots}

\item{shape_var}{svar mapped to shape in feature plots}

\item{group_var}{svar mapped to group in feature plots}

\item{txt_var}{svar mapped to txt   in feature plots}

\item{line}{whether to connect points in feature plot with line (logical)}
}
\description{
Cluster top features (on subgroup profiles)
}
\examples{
\dontrun{
library(magrittr)
if (require(autonomics.data)){
   result_dir <- tempdir() \%T>\% message()
   object <- autonomics.data::stemcomp.proteinratios
   object$subgroup
   contrasts <- c(BM_E = 'BM_E', BM_EM = 'BM_EM', EM_E = 'EM_E')
   object \%>\% autonomics.find::cluster_top_features_on_subgroups(
                   contrasts  = contrasts,
                   result_dir = result_dir,
                   n          = 500)
}

if (require(billing.differentiation.data)){
   library(magrittr)
   result_dir <- tempdir() \%T>\% message()
   object <- billing.differentiation.data::protein.ratios
   contrasts <- c(EM01_0   = 'EM01-EM0.0',
                  EM02_01  = 'EM02-EM01',
                  EM05_02  = 'EM05-EM02',
                  EM15_05  = 'EM15-EM05',
                  EM30_15  = 'EM30-EM15')
   object \%<>\% autonomics.import::filter_samples(subgroup != 'BM0.0')
   object \%>\% cluster_top_features_on_subgroups(
                   contrasts  = contrasts,
                   result_dir = result_dir,
                   n          = 500)
}

if (require(atkin.2014)){
   library(magrittr)
   result_dir <- tempdir() \%T>\% message()
   object <- atkin.2014::soma
   object$subgroup
   contrasts <- c(Diabetic_Control = 'D.t0 - C.t0')
   object \%>\% cluster_top_features_on_subgroups(
                   contrasts = contrasts,
                   result_dir = result_dir)
}

if (require(subramanian.2016)){
library(magrittr)
   result_dir <- tempdir() \%T>\% message()
   object <- subramanian.2016::metabolon
   contrasts <- subramanian.2016::contrasts.metabolon
   object \%>\% cluster_top_features_on_subgroups(
                   contrasts = contrasts,
                   result_dir = result_dir,
                   color_var = 'condition',
                   group_var = 'condition')
}
}

}
