% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_top_definition.R
\name{complete_top_definition}
\alias{complete_top_definition}
\title{Complete top definition}
\usage{
complete_top_definition(top_definition, contrast_name, direction = c("both",
  "neg", "both"))
}
\arguments{
\item{top_definition}{top definition}

\item{contrast_name}{contrast name (string)}

\item{direction}{direction}
}
\value{
completed top definition
}
\description{
Complete top definition
}
\examples{
contrast <- c(P3S = 'P3.S - P3.A')
autonomics.find::complete_top_definition(
   top_definition = "abs(quantile) < 0.05", 
   contrast_name  = names(contrast), 
   direction = 'neg')
}
