% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_top_definition.R
\name{complete_topdef}
\alias{complete_topdef}
\title{Complete top definition}
\usage{
complete_topdef(topdef, contrast_name, direction = c("both", "neg", "both"))
}
\arguments{
\item{topdef}{top definition}

\item{contrast_name}{contrast name (string)}

\item{direction}{direction}
}
\value{
completed top definition
}
\description{
Complete top definition
}
\examples{
contrast <- c(P3S = 'P3.S - P3.A')
autonomics.find::complete_topdef(
   topdef = "abs(quantile) < 0.05", 
   contrast_name  = names(contrast), 
   direction = 'neg')
}
