% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_top_features.R
\name{compute_median_per_subgroup}
\alias{compute_median_per_subgroup}
\title{Compute median per subgroup}
\usage{
compute_median_per_subgroup(object)
}
\arguments{
\item{object}{eSet}
}
\description{
Compute median per subgroup
}
\examples{
require(magrittr)

# STEM CELL COMPARISON
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios
   object \%>\% autonomics.find::compute_median_per_subgroup() \%>\% head()
}
if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon
   object \%>\% autonomics.find::compute_median_per_subgroup()
}
}
