% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains_limma_in_fdata.R
\name{contains_limma_in_fdata}
\alias{contains_limma_in_fdata}
\title{Does eset contain limma results}
\usage{
contains_limma_in_fdata(object)
}
\arguments{
\item{object}{eset}
}
\value{
logical
}
\description{
Does eset contain limma results
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::billing2016
   object \%>\% autonomics.find::contains_limma_in_fdata()
   object \%>\% autonomics.find::add_limma_to_fdata() \%>\% 
                autonomics.find::contains_limma_in_fdata()
}
if (require(billing.differentiation.data)){
   billing.differentiation.data::rna.voomcounts \%>\% 
      contains_limma_in_fdata()
}
}
