% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_subgroups.R
\name{contrast_strings}
\alias{contrast_strings}
\title{Contrast values in character vector}
\usage{
contrast_strings(x, subtractor = " - ", parenthesize = FALSE)
}
\arguments{
\item{x}{character(n)}

\item{subtractor}{character(1)}

\item{parenthesize}{logical(1): whether to parenthesize subtrahend}
}
\value{
character(n-1)
}
\description{
Contrast values in character vector
}
\examples{
require(magrittr)
letters[1:5] \%>\% contrast_strings()
letters[1:5] \%>\% contrast_strings(parenthesize = TRUE)
}
