% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_subgroups.R
\name{contrast_subgroups}
\alias{contrast_subgroups}
\alias{contrast_subgroups.character}
\alias{contrast_subgroups.matrix}
\alias{contrast_subgroups.SummarizedExperiment}
\alias{contrast_contrasts}
\alias{contrast_contrasts.matrix}
\alias{contrast_contrasts.SummarizedExperiment}
\title{Contrast subgroups. Contrast contrasts.}
\usage{
contrast_subgroups(x, ...)

\method{contrast_subgroups}{character}(x)

\method{contrast_subgroups}{matrix}(x)

\method{contrast_subgroups}{SummarizedExperiment}(x)

contrast_contrasts(x, ...)

\method{contrast_contrasts}{matrix}(x)

\method{contrast_contrasts}{SummarizedExperiment}(x)
}
\arguments{
\item{x}{SummarizedExperiment, matrix, or character(.)}
}
\value{
character(.)
}
\description{
Contrast subgroups. Contrast contrasts.
}
\examples{
require(magrittr)

# STEM CELL DIFFERENTIATION
if (require(autonomics.data)){
   x <- autonomics.data::stemdiff.proteinratios
   x \%>\% autonomics.find::contrast_subgroups()
}

# GLUTAMINASE
if (require(autonomics.data)){

   # SummarizedExperiment
   x <- autonomics.data::glutaminase
   x \%>\% contrast_subgroups()
   x \%>\% contrast_contrasts()
   
   # Matrix
   x <- autonomics.data::glutaminase \%>\% autonomics.find::layout_composite_values()
   x \%>\% contrast_subgroups()
   x \%>\% contrast_contrasts()
   
   # Vector
   x <- autonomics.data::glutaminase \%>\% autonomics.find::layout_composite_values() \%>\% magrittr::extract(1,)
   x \%>\% contrast_subgroups()
}

# ATKIN
if (require(atkin.2014)){

   # SummarizedExperiment
   x <- atkin.2014::soma.2018
   x \%>\% contrast_subgroups()
   x \%>\% contrast_contrasts()
           
 # matrix 
   x <- atkin.2014::soma.2018 \%>\% autonomics.find::layout_composite_values()
   x \%>\% contrast_subgroups()
   x \%>\% contrast_contrasts()
   
 # character vector
   x <- atkin.2014::soma.2018 \%>\% autonomics.find::layout_composite_values() \%>\% 
                       magrittr::extract(1, )
   x \%>\% contrast_subgroups()
}

}
