% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_arrow_diagram.R
\name{create_arrow_diagram}
\alias{create_arrow_diagram}
\title{Create arrow diagram}
\usage{
create_arrow_diagram(object, topdef)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{topdef}{character(1): top definition}
}
\value{
character(1)
}
\description{
Create arrow diagram
}
\examples{
if (require(autonomics.data)){
   object <- autonomics.data::stemdiff.proteinratios
   topdef <- 'fdr < 0.05'
   cat(create_arrow_diagram(object, topdef))
}
}
