% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_features.R
\name{create_cluster_contrasts}
\alias{create_cluster_contrasts}
\title{Create contrasts for t-based clustering}
\usage{
create_cluster_contrasts(object)
}
\description{
Create contrasts for t-based clustering
}
