% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_arrow_diagram.R
\name{create_grid_arrow_plot}
\alias{create_grid_arrow_plot}
\title{Create grid arrow plot}
\usage{
create_grid_arrow_plot(object, color_var = object \%>\%
  autonomics.plot::default_color_var(), color_values = object \%>\%
  autonomics.plot::default_color_values(color_var))
}
\description{
Create grid arrow plot
}
\examples{
object <- subramanian.2016::metabolon
}
