% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_contrasts}
\alias{default_contrasts}
\title{Default contrasts}
\usage{
default_contrasts(object)
}
\arguments{
\item{object}{eset}
}
\value{
contrast definitions
}
\description{
Default contrasts
}
\examples{
require(magrittr)

# STEM CELL COMPARISON
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios 
   object \%>\% autonomics.find::default_contrasts()
}

# STEM CELL DIFFERENTIATION
if (require(billing.differentiation.data)){
   billing.differentiation.data::rna.voomcounts \%>\%
      autonomics.find::default_contrasts() \%>\% data.frame()
}

# GLUTAMINASE
if (require(autonomics.data)){
   object <- autonomics.data::glutaminase
   object \%>\% autonomics.import::filter_samples(TREATMENT != 0)  \%>\%
              autonomics.find::default_contrasts()               \%>\% 
              data.frame()
}

if (require(subramanian.2016)){
   subramanian.2016::metabolon \%>\%
      autonomics.find::default_contrasts() \%>\%
      data.frame()
}
# if (require(atkin.2014)){
#    atkin.2014::metabolon \%>\%
#      autonomics.find::default_contrasts() \%>\%
#      data.frame()
# }
if (require(graumann.lfq)){
   graumann.lfq::lfq.intensities \%>\%
      autonomics.find::default_contrasts() \%>\%
      data.frame()
}
}
