% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_nplot}
\alias{default_nplot}
\title{Default nplot}
\usage{
default_nplot(object)
}
\arguments{
\item{object}{eset}
}
\value{
number of top features to be plotted
}
\description{
Default nplot
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   billing.differentiation.data::rna.voomcounts \%>\% 
      autonomics.find::default_nplot()
}
}
