% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_top_definition}
\alias{default_top_definition}
\title{default top_definition}
\usage{
default_top_definition(object)
}
\arguments{
\item{object}{eset}
}
\value{
default value of top_definition
}
\description{
default top_definition
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   billing.differentiation.data::rna.voomcounts \%>\% 
      autonomics.find::default_top_definition()
}
}
