% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_topdef}
\alias{default_topdef}
\title{default topdef}
\usage{
default_topdef(object)
}
\arguments{
\item{object}{eset}
}
\value{
default value of topdef
}
\description{
default topdef
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   billing.differentiation.data::rna.voomcounts \%>\% 
      autonomics.find::default_topdef()
}
}
