% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{direction_to_sign}
\alias{direction_to_sign}
\title{Convert direction into sign}
\usage{
direction_to_sign(direction)
}
\arguments{
\item{direction}{any value in autonomics.find::DIRECTIONS}
}
\value{
sign (character)
}
\description{
Convert direction into sign
}
