% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DIRECTIONS}
\alias{DIRECTIONS}
\title{DIRECTIONS}
\format{An object of class \code{character} of length 3.}
\usage{
DIRECTIONS
}
\description{
Allowed values for diresction argument
}
\keyword{datasets}
