% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliminate_confounders.R
\name{eliminate_confounders}
\alias{eliminate_confounders}
\title{Eliminate confounders}
\usage{
eliminate_confounders(object, confounders = character())
}
\arguments{
\item{object}{exprs object}

\item{confounders}{confounding svars (character)}
}
\value{
object with confounder effects removed
}
\description{
Eliminate confounders
}
\examples{
require(magrittr)
}
