% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_features.R
\name{feature_sample_heatmap}
\alias{feature_sample_heatmap}
\title{Plot heatmap}
\usage{
feature_sample_heatmap(object)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{scale_features}{TRUE or FALSE: whether to scale (i.e. z-score) features}
}
\description{
Plot heatmap
}
\examples{
if (require(autonomics.data)){
    object <- autonomics.data::glutaminase
    object \%<>\% extract(, order(object$subgroup))
    feature_sample_heatmap(object)
}
}
