% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_contrast_features.R
\name{filter_top_features}
\alias{filter_top_features}
\alias{filter_contrast_features}
\title{Filter (and arrange) contrast features}
\usage{
filter_top_features(...)

filter_contrast_features(object, contrast_name, topdef)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{contrast_name}{character(1): contrast name}

\item{topdef}{character(1): top definition.}
}
\description{
Filter (and arrange) contrast features
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   (object <- autonomics.data::stemdiff.proteinratios)
   contrast_name <-  names(autonomics.import::contrastdefs(object))[1]
   topdef <- 'bonf < 0.05'
   (object \%<>\% autonomics.find::filter_contrast_features(contrast_name, topdef))
   object \%>\% autonomics.import::limma()
}
}
