% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_contrast_features.R
\name{filter_n_arrange_top_features}
\alias{filter_n_arrange_top_features}
\alias{filter_n_arrange_contrast_features}
\title{Filter and arrange contrast features}
\usage{
filter_n_arrange_top_features(...)

filter_n_arrange_contrast_features(object, contrast_name, topdef,
  nmax = Inf)
}
\arguments{
\item{...}{for backward compatibility only}

\item{object}{SummarizedExperiment}

\item{contrast_name}{contrast name}

\item{topdef}{top definition}

\item{nmax}{max number of features}
}
\description{
Filter and arrange contrast features
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   (object <- autonomics.data::stemdiff.proteinratios)
   contrast_name <- names(autonomics.import::contrastdefs(object))[1]
   topdef <- 'bonf < 0.05 & effect<0'
   object \%<>\% filter_n_arrange_contrast_features(contrast_name, topdef)
   object \%>\% autonomics.import::limma()
}
}
