% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_n_arrange_top_features.R
\name{filter_n_arrange_top_features}
\alias{filter_n_arrange_top_features}
\title{Filter and arrange top features}
\usage{
filter_n_arrange_top_features(object, contrast_name, topdef, direction,
  nmax = Inf)
}
\arguments{
\item{object}{eset}

\item{contrast_name}{contrast name}

\item{topdef}{top definition}

\item{direction}{'neg' or 'pos'}

\item{nmax}{max number of features}
}
\description{
Filter and arrange top features
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   object <- billing.differentiation.data::rna.voomcounts
   contrast <-  c(EM0.8_0 = 'EM0.8 - EM0.0')
   topdef <- 'bonf < 0.05'
   autonomics.find::filter_n_arrange_top_features(object, names(contrast), topdef, 'neg') \%>\% dim()
   autonomics.find::filter_n_arrange_top_features(object, names(contrast), topdef, 'both') \%>\% dim()
}
}
