% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_n_arrange_top_features.R
\name{filter_significant_features}
\alias{filter_significant_features}
\alias{filter_top_ranked_features}
\title{Filter significant features}
\usage{
filter_significant_features(object)

filter_top_ranked_features(object, n = 1000)
}
\arguments{
\item{object}{eset}

\item{n}{number of top features to select}
}
\value{
filtered eset
}
\description{
Filter features which are significant for at least one contrast
}
\examples{
library(magrittr)
if (require(autonomics.data)){
   autonomics.data::stemcomp.proteinratios               \%>\% 
      autonomics.find::add_limma_to_fdata()   \%>\% 
      autonomics.find::filter_significant_features()
}
if (require(subramanian.2016)){
   subramanian.2016::metabolon \%>\% 
      autonomics.find::add_limma_to_fdata(
         contrasts = subramanian.2016::contrasts.metabolon \%>\% extract(-length(.))) \%>\% 
      autonomics.find::filter_significant_features() \%>\% invisible()
   
   subramanian.2016::exiqon \%>\% 
      autonomics.find::add_limma_to_fdata(
         contrasts  = subramanian.2016::contrasts.exiqon \%>\% extract(-length(.))) \%>\% 
      autonomics.find::filter_significant_features() \%>\% invisible()
   
   subramanian.2016::rnaseq \%>\% 
      autonomics.find::add_limma_to_fdata(
         contrasts = subramanian.2016::contrasts.rnaseq \%>\% extract(-length(.))) \%>\%
      autonomics.find::filter_significant_features() \%>\% 
      autonomics.find::filter_top_ranked_features() \%>\% invisible()
}
}
