% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_n_arrange_top_features.R
\name{filter_top_features}
\alias{filter_top_features}
\title{Filter (and arrange) top features}
\usage{
filter_top_features(object, contrast_name, topdef, direction)
}
\arguments{
\item{object}{eset}

\item{contrast_name}{contrast name}

\item{topdef}{definition of 'top features'.}

\item{direction}{'both', 'neg', or 'pos'}
}
\description{
Filter (and arrange) top features
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   object <- billing.differentiation.data::rna.voomcounts
   contrast <-  c(EM0.8_0 = 'EM0.8 - EM0.0')
   topdef <- 'bonf < 0.05'
   autonomics.find::filter_top_features(object, names(contrast), topdef, 'neg') \%>\% dim()
   autonomics.find::filter_top_features(object, names(contrast), topdef, 'both') \%>\% dim()
}
}
