% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_top_features.R
\name{filter_top_features_per_contrast}
\alias{filter_top_features_per_contrast}
\title{Filter top significant features per contrast}
\usage{
filter_top_features_per_contrast(object, n = 1000)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{n}{number of top features to select.
Actual number of features in returned eset will be lower,
due to overlaps of top features among contrasts.}
}
\description{
Filter top significant features per contrast
}
\examples{
library(magrittr)
if (require(autonomics.data)){
   autonomics.data::stemcomp.proteinratios \%>\% 
   autonomics.find::filter_top_features_per_contrast()
}
}
