% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contrast_subdir.R
\name{get_contrast_subdir}
\alias{get_contrast_subdir}
\title{Get directory with results for particular limma coef}
\usage{
get_contrast_subdir(result_dir, contrast_name)
}
\arguments{
\item{result_dir}{result directory}

\item{contrast_name}{name of limma contrast}
}
\description{
Get directory with results for particular limma coef
}
