% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_contrasts.R
\name{infer_contrast_names}
\alias{infer_contrast_names}
\title{Infer contrast names}
\usage{
infer_contrast_names(object)
}
\arguments{
\item{object}{eset}
}
\value{
character vector
}
\description{
Infer contrast names
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   subramanian.2016::exiqon \%>\%  
      autonomics.find::infer_contrast_names()
}
}
