% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validify_contrasts.R
\name{is_valid_contrast}
\alias{is_valid_contrast}
\title{Is a valid contrast?}
\usage{
is_valid_contrast(contrast, design)
}
\arguments{
\item{contrast}{contrast}

\item{design}{design}
}
\value{
logical
}
\description{
Is a valid contrast?
}
\examples{
if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon
   design <- autonomics.find::create_design_matrix(object)
   contrast <- autonomics.find::default_contrasts(object)[1]
   autonomics.find::is_valid_contrast(contrast, design)
}
}
