% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_subgroups.R
\name{layout_composite_values}
\alias{layout_composite_values}
\title{x <- levels(autonomics.data::glutaminase$subgroup)
x %>% autonomics.support::}
\usage{
layout_composite_values(x)
}
\description{
x <- levels(autonomics.data::glutaminase$subgroup)
x %>% autonomics.support::
}
