% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_subgroups.R
\name{make_ref_contrasts}
\alias{make_ref_contrasts}
\title{Make reference contrasts}
\usage{
make_ref_contrasts(object)
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or EList}
}
\value{
named character vector
}
\description{
Make reference contrasts
}
\examples{
require(magrittr)

# STEMCELL DIFFERENTIATION
if (require(autonomics.data)){
   object <- autonomics.data::stemdiff.proteinratios
   object \%>\% autonomics.find::make_ref_contrasts() \%>\% data.frame()
}

# GLUTAMINASE
if (require(autonomics.data)){
   object <- autonomics.data::glutaminase
   object \%>\% autonomics.find::make_ref_contrasts() \%>\% data.frame()
}

if (require(subramanian.2016)){
   subramanian.2016::metabolon \%>\%  
   autonomics.find::make_ref_contrasts() \%>\% data.frame()
}
}
