% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_contrast_features.R
\name{n_contrast_features}
\alias{n_contrast_features}
\title{Get number of contrast features}
\usage{
n_contrast_features(
  object,
  contrast_name,
  topdef = autonomics.find::default_topdef(object)
)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{contrast_name}{character(1): contrast name}

\item{topdef}{character(1): top definition}

\item{...}{for backward compatibility only}
}
\value{
integer vector with indices of top features
}
\description{
Get number of contrast features
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::stemdiff.proteinratios
   topdef <- 'p < 0.05'
   contrast_name <- object \%>\% autonomics.import::contrastdefs() \%>\% magrittr::extract(1) \%>\% names()
   object \%>\% autonomics.find::n_contrast_features('p<0.05',            contrast_name) \%>\% sum()
   object \%>\% autonomics.find::n_contrast_features('p<0.05 & effect>0', contrast_name) \%>\% sum()
}
}
