% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_confounder_effects.R
\name{plot_confounder_effects}
\alias{plot_confounder_effects}
\title{Plot confounder effects}
\usage{
plot_confounder_effects(object, confounders = character(), fvars,
  color_var = "subgroup")
}
\arguments{
\item{object}{exprs object}

\item{confounders}{confounding svars}

\item{fvars}{fvars}

\item{color_var}{svar mapped to color}
}
\description{
Plot confounder effects
}
