% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contrast_features.R
\name{plot_top_features}
\alias{plot_top_features}
\alias{plot_contrast_features}
\title{Plot contrast features}
\usage{
plot_top_features(...)

plot_contrast_features(
  object,
  contrast = autonomics.find::default_contrasts(object)[1],
  design = autonomics.import::create_design_matrix(object),
  topdef = autonomics.find::default_topdef(object),
  geom = autonomics.plot::default_feature_plots(object)[1],
  nplot = autonomics.find::default_nplot(object),
  fvars = autonomics.plot::default_fvars(object),
  color_var = autonomics.plot::default_color_var(object),
  ...
)
}
\arguments{
\item{...}{passed to autonomics.plot::plot_feature_xxx()}

\item{object}{\code{SummarizedExperiment}}

\item{contrast}{named contrast for which to select the top feature bars}

\item{design}{design matrix}

\item{topdef}{definition of 'top features'.}

\item{geom}{'point', 'boxplot', 'violin', 'bar', 'hbar'}

\item{nplot}{max no of top features to plot}

\item{fvars}{fvars to use in plot}
}
\description{
This function visualizes the differential contrast features
}
\examples{
require(magrittr)

# STEMCELL COMPARISON
  file <- tempfile() \%>\% paste0('.pdf') \%T>\% message()
  if (require(autonomics.data)){
     object <- autonomics.data::stemcomp.proteinratios
     object \%>\% plot_contrast_features(geom = 'hbar')
     object \%>\% plot_contrast_features(geom = 'bar')
     object \%>\% plot_contrast_features(geom = 'point')
     object \%>\% plot_contrast_features(geom = 'violin')
     object \%>\% plot_contrast_features(geom = 'boxplot')
     object \%>\% plot_contrast_features(x = 'subgroup')
  }

# GLUTAMINASE
  if (require(autonomics.data)){
     object <- autonomics.data::glutaminase
     topdef <- 'bonf < 0.05 & rank <= 4'
     object \%>\% plot_contrast_features(topdef = topdef, geom = 'boxplot')
  }

# A somascan eset
if (require(atkin.2014)){
   object  <- atkin.2014::soma
   contrasts <- atkin.2014::contrasts
   object \%>\% plot_contrast_features(contrast = contrasts[1], nplot = 9,
                                geom = 'point', x = 'time', color_var = 'condition',
                                facet_var = 'subject_id', fvars = 'TargetFullName')
}

# A metabolon eset
if (require(subramanian.2016)){
   contrasts <- subramanian.2016::contrasts.metabolon[1]
   object  <- subramanian.2016::metabolon \%>\% add_limma_to_fdata(contrasts = contrasts)
   object \%>\% plot_contrast_features(contrast  = contrasts,
                                geom      = 'boxplot',
                                color_var = 'condition',
                                group_var = 'condition',
                                line      = TRUE,
                                nplot     = 4)
}

if (require(billing.differentiation.data)){
   object <- billing.differentiation.data::rna.voomcounts
   contrast <- billing.differentiation.data::contrasts[1]
   object \%>\% autonomics.find::plot_contrast_features(contrast = contrast, 
                                                 nplot    = 4)
}


}
\author{
Aditya Bhagwat
}
