% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contrast_venns.R
\name{plot_contrast_venns}
\alias{plot_contrast_venns}
\title{Plot contrast venns}
\usage{
plot_contrast_venns(object, contrast_names = object \%>\%
  autonomics.find::infer_contrast_names(),
  topdef = autonomics.find::default_topdef(object), title = topdef,
  euler = FALSE, file = NULL, ...)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{contrast_names}{character vector}

\item{topdef}{top definition}

\item{title}{plot title}

\item{euler}{logical(1): whether to plot euler rather than venn diagram}

\item{file}{file to which print to}

\item{...}{autonomics.plot::plot_venn(...)}
}
\value{
the vennlist used for venn digram plotting
}
\description{
Plot contrast venns
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   object <- subramanian.2016::exiqon
   object \%>\% autonomics.find::plot_contrast_venns()
}
}
