% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastogram.R
\name{plot_contrastogram}
\alias{plot_contrastogram}
\title{Plot contrastogram}
\usage{
plot_contrastogram(
  object,
  directed = TRUE,
  subgroup_colors = default_color_values2(object),
  sparse = FALSE
)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{directed}{TRUE or FALSE: whether to distinguish up and downregulations}

\item{subgroup_colors}{named color vector (names = subgroups)}
}
\description{
Plot contrastogram
}
\examples{
if (require(autonomics.data)){
    object <- autonomics.data::glutaminase
    plot_contrastogram(object)
    plot_contrastogram(object, directed = FALSE)
}
}
