% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contrast_features.R
\name{plot_top_features_all_contrasts}
\alias{plot_top_features_all_contrasts}
\alias{plot_per_contrast_features}
\title{Plot per contrast features for all contrasts}
\usage{
plot_top_features_all_contrasts(...)

plot_per_contrast_features(object,
  contrasts = autonomics.find::default_contrasts(object),
  design = autonomics.import::create_design_matrix(object),
  topdef = autonomics.find::default_topdef(object),
  geom = autonomics.plot::default_feature_plots(object), result_dir, ...)
}
\arguments{
\item{...}{passed to autonomics.plot::plot_features}

\item{object}{SummarizedExperiment}

\item{contrasts}{named contrast vector}

\item{design}{design matrix}

\item{geom}{which feature plots to be created?}

\item{result_dir}{directory where to store results}
}
\description{
Plot per contrast features for all contrasts
}
\examples{
require(magrittr)
if (require(atkin.2014)){
   object <- atkin.2014::soma
   contrasts <- atkin.2014::contrasts
   result_dir <- tempdir() \%T>\% message()
   object \%>\% plot_per_contrast_features(
                   contrasts = contrasts[1], 
                    result_dir = result_dir)
   object \%>\% plot_per_contrast_features(
                   contrasts  = contrasts[1], 
                   result_dir = result_dir,
                   nplot      = 9, 
                   geom       = 'point', 
                   x          = 'time',
                   color_var  = 'condition',
                   facet_var  = 'subject_id',
                   fvars      = 'TargetFullName')
}
# GLUTAMINASE
if (require(autonomics.data)){
  object <- autonomics.data::glutaminase
  result_dir <- tempdir() \%T>\% message()
  object \%>\% plot_per_contrast_features(
                contrasts  = autonomics.import::contrastdefs(object)[1:2],
                topdef     = 'fdr < 0.05',
                geom       = 'boxplot',
                result_dir = result_dir)
}
}
