% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top_venns.R
\name{plot_top_venns}
\alias{plot_top_venns}
\title{Plot venns showing top features}
\usage{
plot_top_venns(object, direction,
  topdef = autonomics.find::default_topdef(object), file = NULL)
}
\arguments{
\item{object}{exprs object}

\item{direction}{'neg' or 'pos'}

\item{topdef}{top definition}

\item{file}{file to which print to}
}
\value{
the vennlist used for venn digram plotting
}
\description{
Plot venns showing top features
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   object <- subramanian.2016::exiqon
   object \%>\% autonomics.find::plot_top_venns('neg')
}
}
