% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_limma2.R
\name{run_limma}
\alias{run_limma}
\title{Run limma}
\usage{
run_limma(object, contrasts = autonomics.find::default_contrasts(object),
  design = autonomics.find::create_design_matrix(object))
}
\arguments{
\item{object}{eset}

\item{contrasts}{contrast vector, preferably named (automatically generated names are not always intuitive)}

\item{design}{design matrix}
}
\value{
list with four components: rank, fc, p, fdr, bonf
}
\description{
Compare contrasts of object$subgroup levels
}
\examples{
require(magrittr)
#if (require(atkin.2014)){
#   results <- atkin.2014::soma  \%>\%  autonomics.find::run_limma()
#}

# STEM CELL DIFFERENTIATION
if (require(autonomics.data)){
   object <- autonomics.data::stemdiff.proteinratios
   contrasts <- autonomics.find::make_ref_contrasts(object)
   object \%>\% autonomics.find::run_limma(contrasts[1:2]) \%>\% 
              extract2('bonf')       \%>\%
              extract(, 'EM01_EM00') \%>\% 
              is_less_than(0.05)     \%>\% 
              sum(na.rm=TRUE)

}
if (require(billing.differentiation.data)){
   contrasts <- billing.differentiation.data::contrasts

   billing.differentiation.data::rna.counts \%>\%
      autonomics.import::logcpm() \%>\%
      autonomics.find::run_limma(contrasts) \%>\% extract2('bonf') \%>\%
      extract(, 'EM01_0') \%>\% is_less_than(0.05) \%>\% sum(na.rm=TRUE)

   billing.differentiation.data::rna.counts \%>\%
      autonomics.import::voom_transform() \%>\%
      autonomics.find::run_limma(contrasts) \%>\% extract2('bonf') \%>\%
      extract(, 'EM01_0') \%>\% is_less_than(0.05) \%>\% sum(na.rm=TRUE)
}
if (require(subramanian.2016)){
   dir <- system.file('extdata/rnaseq', package = 'subramanian.2016')
   object <- autonomics.import::load_rnaseq_counts(dir, sample_file = NULL) \%>\% 
             autonomics.import::voom_transform()
   results <- object \%>\% autonomics.find::run_limma()
   results$p[1:3,1:3]
}

}
