% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_fvars_and_filter_samples.R
\name{select_fvars_and_filter_samples}
\alias{select_fvars_and_filter_samples}
\title{Select portion of eset relevant to contrasts}
\usage{
select_fvars_and_filter_samples(object, design = create_design_matrix(object),
  contrasts)
}
\arguments{
\item{object}{eSet}

\item{design}{design matrix}

\item{contrasts}{contrasts of subgroup levels (named string)}
}
\value{
Filtered eSet
}
\description{
Select portion of eset relevant to contrasts
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   object <- autonomics.data::billing2016 \%>\% autonomics.find::add_limma_to_fdata()
   contrasts <- object \%>\% autonomics.find::default_contrasts() \%>\% magrittr::extract(1)
   object \%>\% autonomics.find::select_fvars_and_filter_samples(contrasts = contrasts)
}
if (require(billing.differentiation.data)){
   object <- billing.differentiation.data::rna.voomcounts
   object \%>\% dim()
   contrasts <-  c(EM0.8_0 = 'EM0.8 - EM0.0')
   object \%>\% autonomics.find::select_fvars_and_filter_samples(contrasts = contrasts) \%>\% 
                dim()
}
}
