% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validify_contrasts.R
\name{select_valid_contrasts}
\alias{select_valid_contrasts}
\title{Select valid contrasts}
\usage{
select_valid_contrasts(contrasts, design, verbose = TRUE)
}
\arguments{
\item{contrasts}{vector with contrast definitions}

\item{design}{design matrix}

\item{verbose}{whether or not to report}
}
\value{
subset of contrasts
}
\description{
Select valid contrasts
}
