% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{validify_contrast_names}
\alias{validify_contrast_names}
\title{Validify contrast names}
\usage{
validify_contrast_names(x)
}
\arguments{
\item{x}{vector with contrast names}
}
\value{
vector with valid names
}
\description{
Validify contrast names
}
