% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_features.R
\name{write_all_features}
\alias{write_all_features}
\title{Write all features to file}
\usage{
write_all_features(object, result_dir)
}
\arguments{
\item{object}{eset}

\item{result_dir}{dir}
}
\description{
Write all features to file
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::ALL
   result_dir <- tempdir()
   autonomics.find::write_all_features(object, result_dir)
   dir(result_dir)
   unlink(result_dir, recursive = TRUE)
}
}
