% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_features.R
\name{write_features}
\alias{write_features}
\title{Write feature tables to file}
\usage{
write_features(object, design = create_design_matrix(object), contrasts,
  top_definition, result_dir)
}
\arguments{
\item{object}{eset}

\item{design}{design matrix}

\item{contrasts}{A character vector of model contrasts to be written.}

\item{top_definition}{top definition}

\item{result_dir}{dir where to save results}
}
\description{
Write feature tables to file
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object  <- autonomics.data::billing2016 \%>\% 
                autonomics.find::add_limma_to_fdata()
   contrasts <- autonomics.find::default_contrasts(object)
   top_definition <- object \%>\% autonomics.find::default_top_definition()
   result_dir <- tempdir()
   autonomics.find::write_features(object, contrasts = contrasts,
                                   top_definition = top_definition, result_dir = result_dir)
}
}
