% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_features.R
\name{write_top_features}
\alias{write_top_features}
\title{Write top features for specified contrast}
\usage{
write_top_features(object, design = create_design_matrix(object), contrast,
  direction = autonomics.find::DIRECTIONS[1],
  topdef = default_topdef(object), result_dir = NULL)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{design}{design matrix}

\item{contrast}{see \code{analyze_eset}}

\item{direction}{any value in autonomics.find::DIRECTIONS}

\item{topdef}{top definition}

\item{result_dir}{directory}
}
\description{
Write top features for specified contrast
}
\examples{
require(magrittr)

# STEM CELL COMPARISON
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios
   result_dir <- tempdir() \%T>\% message()
   object \%>\% autonomics.find::write_top_features(
                 contrast       = c(BM_E = 'BM_E'),
                 direction      = 'neg', 
                 topdef = 'fdr < 0.5', 
                 result_dir     = result_dir)
}

# STEM CELL DIFFERENTIATION
if (require(billing.differentiation.data)){
   object <- billing.differentiation.data::rna.voomcounts
   result_dir <- tempdir() \%T>\% message()
   object \%>\% autonomics.find::write_top_features(
      contrast = billing.differentiation.data::contrasts[1], result_dir = result_dir)
   object \%>\% autonomics.find::write_top_features(
      contrast   = billing.differentiation.data::contrasts[1],
      direction  = 'both', 
      result_dir = result_dir)
}
}
