% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastogram.R
\name{aggregate_contrasts}
\alias{aggregate_contrasts}
\title{Aggregate time contrasts across conc (or vice versa)}
\usage{
aggregate_contrasts(contrastmat, dim)
}
\arguments{
\item{contrastmat}{contrast matrix}
}
\description{
Aggregate time contrasts across conc (or vice versa)
}
