% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_n_arrange_top_features.R
\name{are_top_features}
\alias{are_top_features}
\alias{get_top_features}
\title{Which are the top features?}
\usage{
are_top_features(object, topdef = autonomics.find::default_topdef(object),
  contrast_name)

get_top_features(object, topdef = autonomics.find::default_topdef(object),
  contrast_name)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{topdef}{top definition}

\item{contrast_name}{contrast name (string)}
}
\value{
integer vector with indices of top features
}
\description{
Which features fulfill top definition and direction for specified contrast?
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::stemdiff.proteinratios
   topdef <- 'p < 0.05'
   contrast_name <- object \%>\% autonomics.import::contrastdefs() \%>\% magrittr::extract(1) \%>\% names()
   object \%>\% autonomics.find::are_top_features('p<0.05',            contrast_name) \%>\% sum()
   object \%>\% autonomics.find::are_top_features('p<0.05 & effect>0', contrast_name) \%>\% sum()
}
}
