% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_features.R
\name{cluster_features}
\alias{cluster_features}
\title{Cluster features}
\usage{
cluster_features(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\description{
Cluster features
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   object <- subramanian.2016::exiqon
   object \%>\% autonomics.plot::plot_overlayed_sample_distributions()
}
}
