% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_contrast_features.R
\name{filter_significant_features}
\alias{filter_significant_features}
\alias{filter_top_ranked_features}
\title{Filter significant features}
\usage{
filter_significant_features(object)

filter_top_ranked_features(object, n = 1000)
}
\arguments{
\item{object}{eset}

\item{n}{number of top features to select}
}
\value{
filtered eset
}
\description{
Filter features which are significant for at least one contrast
}
\examples{
library(magrittr)
if (require(autonomics.data)){
   autonomics.data::stemcomp.proteinratios \%>\% autonomics.find::filter_significant_features()
}
if (require(subramanian.2016)){
   subramanian.2016::metabolon \%>\% autonomics.find::filter_significant_features()
   subramanian.2016::exiqon    \%>\% autonomics.find::filter_significant_features()
   subramanian.2016::rnaseq    \%>\% autonomics.find::filter_significant_features() 
}
}
