% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_n_arrange_top_features.R
\name{filter_top_features}
\alias{filter_top_features}
\title{Filter (and arrange) top features}
\usage{
filter_top_features(object, contrast_name, topdef)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{contrast_name}{contrast name}

\item{topdef}{definition of 'top features'.}

\item{direction}{'both', 'neg', or 'pos'}
}
\description{
Filter (and arrange) top features
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   (object <- autonomics.data::stemdiff.proteinratios)
   contrast_name <-  names(autonomics.import::contrastdefs(object))[1]
   topdef <- 'bonf < 0.05'
   (object \%<>\% autonomics.find::filter_top_features(contrast_name, topdef))
   object \%>\% autonomics.import::limma()
}
}
