% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_contrasts.R
\name{make_ref_contrasts}
\alias{make_ref_contrasts}
\title{Make reference contrasts}
\usage{
make_ref_contrasts(object)
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or EList}
}
\value{
named character vector
}
\description{
Make reference contrasts
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   billing.differentiation.data::protein.ratios \%>\% 
      autonomics.find::make_ref_contrasts() \%>\% data.frame()
}
if (require(subramanian.2016)){
   subramanian.2016::metabolon \%>\%  
      autonomics.find::make_ref_contrasts() \%>\% data.frame()
}
if (require(halama.2016)){
   halama.2016::cell.metabolites \%>\% autonomics.find::make_ref_contrasts() \%>\% data.frame()
   halama.2016::cell.metabolites \%>\% 
      autonomics.import::filter_samples(GROUP_DESCRIPTION != 'Control') \%>\% 
      autonomics.find::make_ref_contrasts() \%>\% 
      data.frame()
}
}
