% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top_features.R
\name{plot_top_features_all_contrasts}
\alias{plot_top_features_all_contrasts}
\title{Plot top features}
\usage{
plot_top_features_all_contrasts(object,
  design = autonomics.import::create_design_matrix(object),
  contrasts = autonomics.find::default_contrasts(object),
  direction = c("neg", "pos"),
  topdef = autonomics.find::default_topdef(object), result_dir,
  geom = autonomics.plot::default_feature_plots(object), ...)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{design}{design matrix}

\item{contrasts}{named contrast vector}

\item{direction}{subset of c('neg', 'pos', 'both')}

\item{result_dir}{directory where to store results}

\item{geom}{which feature plots to be created?}

\item{...}{passed to autonomics.plot::plot_features}
}
\description{
Plot top features per contrast
}
\examples{
require(magrittr)
if (require(atkin.2014)){
   object <- atkin.2014::soma
   contrasts <- atkin.2014::contrasts
   result_dir <- tempdir() \%T>\% message()
   object \%>\% plot_top_features_all_contrasts(
                   contrasts = contrasts[1], 
                    result_dir = result_dir)
   object \%>\% plot_top_features_all_contrasts(
                   contrasts  = contrasts[1], 
                   result_dir = result_dir,
                   nplot      = 9, 
                   geom       = 'point', 
                   x          = 'time',
                   color_var  = 'condition',
                   facet_var  = 'subject_id',
                   fvars      = 'TargetFullName')
}
# GLUTAMINASE
if (require(autonomics.data)){
  object <- autonomics.data::glutaminase
  result_dir <- tempdir() \%T>\% message()
  object \%>\% plot_top_features_all_contrasts(
                contrasts      = autonomics.import::contrastdefs(.)[1:2],
                topdef = 'fdr < 0.05',
                x              = 'TIME_POINT',
                geom           = 'boxplot',
                result_dir     = result_dir)
}
}
