% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_fvars_and_filter_samples.R
\name{select_relevant_fvars}
\alias{select_relevant_fvars}
\title{Select relevant fvars}
\usage{
select_relevant_fvars(object, contrast_names)
}
\arguments{
\item{object}{eset}

\item{contrast_names}{contrast names}
}
\value{
object
}
\description{
Select relevant fvars
}
\examples{
require(magrittr)

# STEM CELL COMPARISON
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios
   contrasts <- object \%>\% autonomics.find::default_contrasts()
   object \%<>\% autonomics.find::add_limma_to_fdata(contrasts = contrasts)
   object \%>\% autonomics.import::fvars() \%>\% length()
   object \%>\% autonomics.find::select_relevant_fvars(names(contrasts[1:2])) \%>\% 
              autonomics.import::fvars() \%>\% length()
}
if (require(billing.differentiation.data)){
   object <- billing.differentiation.data::rna.voomcounts
   object \%>\% autonomics.import::fdata() \%>\% dim()
   contrasts <-  c(EM0.8_0 = 'EM0.8 - EM0.0')
   object \%>\% autonomics.find::select_relevant_fvars(names(contrasts)) \%>\% 
                autonomics.import::fdata() \%>\% dim()
}
# This shows why it is useful to have contrast_names (only) as 
# inputs rather than contrasts
if (require(subramanian.2016)){
   object <- subramanian.2016::exiqon
   (contrast_names <- object \%>\% autonomics.find::infer_contrast_names())
   contrast_names \%<>\% magrittr::extract(-c(2,3))
   object \%>\% autonomics.find::select_relevant_fvars(contrast_names) \%>\% 
                autonomics.import::fdata() \%>\% dim()
}
}
