% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validify_contrasts.R
\name{validify_contrasts}
\alias{validify_contrasts}
\title{Validy subgroups and contrasts}
\usage{
validify_contrasts(contrasts, design)
}
\arguments{
\item{contrasts}{contrasts vector}

\item{design}{design matrix}
}
\value{
validified contrasts
}
\description{
Validy subgroups and contrasts
}
\examples{
require(magrittr)

# STEM CELL COMPARISON
if (require(autonomics.data)){
   object  <- autonomics.data::stemcomp.proteinratios
   design <- autonomics.import::create_design_matrix(object)
   contrasts <- autonomics.find::make_ref_contrasts(object)
   contrasts \%>\% autonomics.find::validify_contrasts(design)
}
}
