% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_features.R
\name{write_features}
\alias{write_features}
\title{Write feature tables to file}
\usage{
write_features(object, design = autonomics.find::create_design_matrix(object),
  contrasts = autonomics.find::default_contrasts(object),
  direction = c("neg", "pos"), topdef, result_dir)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{design}{design matrix}

\item{contrasts}{A character vector of model contrasts to be written.}

\item{direction}{'neg', 'pos', 'both'}

\item{topdef}{top definition}

\item{result_dir}{dir where to save results}
}
\description{
Write feature tables to file
}
\examples{
require(magrittr)

# STEM CELL COMPARISON
if (require(autonomics.data)){
   object  <- autonomics.data::stemcomp.proteinratios
   result_dir <- tempdir() \%T>\% message()
   object \%>\% autonomics.find::write_features(
                 contrasts      =  c(BM_E = 'BM_E', BM_EM = 'BM_EM', EM_E = 'EM_E'),
                 topdef = 'fdr < 0.05', 
                 result_dir     =  result_dir)
}
}
