% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_features.R
\name{write_top_features}
\alias{write_top_features}
\title{Write top features for specified contrast}
\usage{
write_top_features(object, design = create_design_matrix(object), contrast,
  direction = autonomics.find::DIRECTIONS[1],
  top_definition = default_top_definition(object), result_dir = NULL)
}
\arguments{
\item{object}{eset}

\item{design}{design matrix}

\item{contrast}{see \code{analyze_eset}}

\item{direction}{any value in autonomics.find::DIRECTIONS}

\item{top_definition}{top definition}

\item{result_dir}{directory}
}
\description{
Write top features for specified contrast
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   object <- autonomics.data::billing2016
   contrast <- object \%>\% autonomics.find::default_contrasts() \%>\% extract(1)
   object \%<>\% autonomics.find::add_limma_to_fdata(contrasts = contrast)
   direction <- 'neg'
   top_definition <- autonomics.find::default_top_definition(object)
   result_dir = tempdir()
   autonomics.find::write_top_features(object, contrast = contrast,
      direction = direction, top_definition = top_definition, result_dir = result_dir)
}
if (require(billing.differentiation.data)){
   require(magrittr)
   object <- billing.differentiation.data::rna.voomcounts
   result_dir <- tempdir() \%T>\% message()
   object \%>\% autonomics.find::write_top_features(
      contrast = billing.differentiation.data::contrasts[1], result_dir = result_dir)
   object \%>\% autonomics.find::write_top_features(
      contrast   = billing.differentiation.data::contrasts[1],
      direction  = 'both', 
      result_dir = result_dir)
}
}
